+++ tests/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vanilla-extract-private/tests
 
+## 0.0.44
+
+### Patch Changes
+
+- Updated dependencies [[`fd5fac5`](https://github.com/vanilla-extract-css/vanilla-extract/commit/fd5fac56573c65f3cd9326cfaa1d2835a7212bcb), [`fd5fac5`](https://github.com/vanilla-extract-css/vanilla-extract/commit/fd5fac56573c65f3cd9326cfaa1d2835a7212bcb)]:
+  - @vanilla-extract/recipes@0.5.0
+  - @vanilla-extract-private/test-helpers@0.0.43
+
 ## 0.0.43
 
 ### Patch Changes